Add GPS while onPause fix
My GPSService was sending broadcast while the app was paused which was resulting in the broadcast not getting to the activities. I added a flag in service which blocks sending broadcast during the pause state of the activity. So now - when the user gets to the interesting place while the app is paused (device is blocked i.e.) - service will wait and inform the activity about that event when it's resumed resulting in displaying info about that new place