Add bugfix
I had a problem with restarting the walk while being in the range of the starting point. GPSService remembered last posistion (which was a starting point) and after restart wasn't sending a broadcast about being in this point because it was an old one. The only thing I had to do was to set lastPlace to null while restarting the thread ;)